feat: images map in redux store with listener
